Faundatrek/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==11.3.0
gunicorn==21.2.0
whitenoise==6.6.0
